8-4-22
OPT 101, 103-111

+++ Assets/001 - Opticlash/002 - Scripts/007 - ScriptableObject/003 - Weapon/CustomWeaponData.cs
@@ -7,6 +7,6 @@ public class CustomWeaponData : ScriptableObject
 {
     [field: SerializeField] public string WeaponInstanceID { get; set; }
     [field: SerializeField] public WeaponData BaseWeaponData { get; set; }
-    [field: SerializeField] public float BonusDamage { get; set; }
+    [field: SerializeField] public int BonusDamage { get; set; }
 
 }
